move fetch times into componentDidMount
